Rimuovi handler legacy e artefatti
